Temporary fix for getting bad data
At the time of the fix, the API sends garbage data. There's
near-duplicate data for some (but not all) entries with just
extra ranks added randomly.

To combat this for now, I filter the initial list to unique
entries by entry id and throw out everything else.

This has the possibility to be wrong if more trash data comes in
later but for now, it seems to work. I'll fix it again when the next
update happens, if it breaks.